requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.14
openpyxl==3.1.2
pytest==8.0.2
python-multipart==0.0.9
ortools==9.8.3296; platform_system != "Darwin"
ortools==9.12.4544; platform_system == "Darwin"
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
reportlab==4.2.0
